requirements.txt
flask
twilio
ghana-nlp
pydub
requests
python-dotenv
